+++ src/components/Library/Sections/WatchLater.tsx
@@ -1,4 +1,4 @@
-import VideoCard from '@components/Common/VideoCard'
+import VideoCard from '@components/CommonT/VideoCard'
 import { NoDataFound } from '@components/UIElements/NoDataFound'
 import useAppStore from '@lib/store'
 import { WATCH_LATER_LIBRARY } from '@utils/url-path'
