+++ src/components/Settings/index.tsx
@@ -1,5 +1,5 @@
-import Layout from '@components/Common/Layout'
-import MetaTags from '@components/Common/MetaTags'
+import Layout from '@components/CommonT/Layout'
+import MetaTags from '@components/CommonT/MetaTags'
 import React from 'react'
 
 const Settings = () => {
